monitoring/requirements.txt
kafka-python
prometheus_client
requests
